backend/pyproject.toml
[project]
name = "aimarket-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "dependency-injector>=4.43",
    "structlog>=24.4",
    "python-jose[cryptography]>=3.3",
    "langgraph>=0.2",
    "langgraph-checkpoint-postgres>=2.0",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "tavily-python>=0.5",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "psycopg[binary,pool]>=3.2",
    "sse-starlette>=2.1",
    "langfuse>=4.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
